python/raspi_weather_app/k8s_weather/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
datadog==0.52.1
idna==3.10
mysql-connector-python==9.4.0
numpy==1.26.4
olefile==0.46
pillow==10.2.0
# python-apt==2.7.7+ubuntu5
requests==2.32.5
RTIMULib==7.2.1
sense-hat==2.2.0
setuptools==68.1.2
urllib3==2.5.0
wheel==0.42.0